move core into packages
