+++ packages/browser/src/browser.all.ts
@@ -1,6 +1,6 @@
-import { i18nString } from '../../../src/core/util.string'
-import { i18nObject } from '../../../src/core/util.object'
-import { i18n } from '../../../src/core/util.instance'
+import { i18nString } from '../../core/src/util.string'
+import { i18nObject } from '../../core/src/util.object'
+import { i18n } from '../../core/src/util.instance'
 // import { i18nLoader, i18nLoaderAsync } from '../../src/core/util.loader'
 
 //@ts-ignore
