+++ packages/browser/src/browser.loader.async.ts
@@ -1,3 +1,3 @@
-import { i18nObjectLoaderAsync } from '../../../src/core/util.loader'
+import { i18nObjectLoaderAsync } from '../../core/src/util.loader'
 //@ts-ignore
 window.i18nObjectLoaderAsync = i18nObjectLoaderAsync
