+++ packages/browser/src/browser.loader.ts
@@ -1,3 +1,3 @@
-import { i18nObjectLoader } from '../../../src/core/util.loader'
+import { i18nObjectLoader } from '../../core/src/util.loader'
 //@ts-ignore
 window.i18nObjectLoader = i18nObjectLoader
