+++ packages/browser/src/browser.object.ts
@@ -1,3 +1,3 @@
-import { i18nObject } from '../../../src/core/util.object'
+import { i18nObject } from '../../core/src/util.object'
 //@ts-ignore
 window.i18nObject = i18nObject
