+++ packages/browser/src/browser.ts
@@ -1,3 +1,3 @@
-import { i18n } from '../../../src/core/util.instance'
+import { i18n } from '../../core/src/util.instance'
 //@ts-ignore
 window.i18n = i18n
