+++ packages/core/src/util.string.ts
