+++ src/adapters/svelte.ts
@@ -1,14 +1,14 @@
-import type { BaseFormatters, BaseTranslation, TranslationFunctions } from '../core/core'
+import type { BaseFormatters, BaseTranslation, TranslationFunctions } from '../../packages/core/src/core'
 import type { Readable, Writable } from 'svelte/store'
 import { derived, writable } from 'svelte/store'
-import { i18nObject } from '../core/util.object'
+import { i18nObject } from '../../packages/core/src/util.object'
 import {
 	FormattersInitializer,
 	AsyncFormattersInitializer,
 	TranslationLoader,
 	TranslationLoaderAsync,
-} from '../core/util.loader'
-import { getFallbackProxy } from '../core/core-utils'
+} from '../../packages/core/src/util.loader'
+import { getFallbackProxy } from '../../packages/core/src/core-utils'
 
 // --------------------------------------------------------------------------------------------------------------------
 // types --------------------------------------------------------------------------------------------------------------
