+++ src/generator/generator-util.ts
@@ -1,4 +1,4 @@
-import type { Arguments, Locale } from '../core/core'
+import type { Arguments, Locale } from '../../packages/core/src/core'
 
 export const getPermutations = <T>(rest: T[], permutatedArray: T[] = []): T[][] => {
 	if (rest.length === 0) {
