+++ src/index.ts
@@ -1,8 +1,8 @@
-export { BaseTranslation } from './core/core'
+export { BaseTranslation } from '../packages/core/src/core'
 
-export { LocaleTranslations, FormattersInitializer, AsyncFormattersInitializer } from './core/util.loader'
+export { LocaleTranslations, FormattersInitializer, AsyncFormattersInitializer } from '../packages/core/src/util.loader'
 
-export { i18nString } from './core/util.string'
-export { i18nObject } from './core/util.object'
-export { i18nObjectLoader, i18nObjectLoaderAsync } from './core/util.loader'
-export { i18n } from './core/util.instance'
+export { i18nString } from '../packages/core/src/util.string'
+export { i18nObject } from '../packages/core/src/util.object'
+export { i18nObjectLoader, i18nObjectLoaderAsync } from '../packages/core/src/util.loader'
+export { i18n } from '../packages/core/src/util.instance'
