+++ src/rollup-plugin-typesafe-i18n-optimizer.ts
@@ -11,8 +11,8 @@ import type {
 import { GeneratorConfig, getConfigWithDefaultValues } from './generator/generator'
 import { createFilter } from '@rollup/pluginutils'
 import { walk } from 'estree-walker'
-import { partsAsStringWithoutTypes } from './core/core-utils'
-import { parseRawText } from './core/parser'
+import { partsAsStringWithoutTypes } from '../packages/core/src/core-utils'
+import { parseRawText } from '../packages/core/src/parser'
 import { generate } from 'astring'
 import sourceMap from 'source-map'
